ai-resume-linkedin-analyzer/requirements.txt
# Web Framework
streamlit==1.31.0
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Web Scraping
selenium==4.17.2
playwright==1.41.2
beautifulsoup4==4.12.3
lxml==5.1.0
webdriver-manager==4.0.1

# Resume Parsing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pytesseract==0.3.10
pdf2image==1.17.0
Pillow==10.2.0

# Natural Language Processing
spacy==3.7.2
transformers==4.37.2
torch==2.1.2
sentence-transformers==2.3.1
nltk==3.8.1

# AI & LLM
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.18.0

# Vector Database & Embeddings
faiss-cpu==1.7.4
langchain==0.1.4
chromadb==0.4.22

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching
redis==5.0.1
python-memcached==1.59

# Data Processing
pandas==2.2.0
numpy==1.26.3
scikit-learn==1.4.0

# Utilities
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Monitoring & Logging
loguru==0.7.2
prometheus-client==0.19.0
sentry-sdk==1.40.0

# HTTP Clients
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Task Queue (Optional)
celery==5.3.6
flower==2.0.1

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.3
pytest-mock==3.12.0
faker==22.6.0
freezegun==1.4.0

# Code Quality
black==24.1.1
isort==5.13.2
flake8==7.0.0
pylint==3.0.3
mypy==1.8.0
pre-commit==3.6.0

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==2.0.0
mkdocs==1.5.3
mkdocs-material==9.5.7

# Data Validation
jsonschema==4.21.0
cerberus==1.3.5

# Date & Time
python-dateutil==2.8.2
arrow==1.3.0
pytz==2024.1

# File Handling
openpyxl==3.1.2
xlrd==2.0.1
chardet==5.2.0

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Rate Limiting
slowapi==0.1.9
limits==3.7.0

# Configuration
pyyaml==6.0.1
python-decouple==3.8
hydra-core==1.3.2

# Progress Bars
tqdm==4.66.1
rich==13.7.0

# Retry Logic
tenacity==8.2.3
backoff==2.2.1

# Environment Detection
python-dotenv==1.0.1